pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gleitzeit"
version = "0.0.4"
description = "Protocol-based workflow orchestration system for LLM and task automation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Leif Markthaler"},
]
keywords = ["workflow", "orchestration", "automation", "distributed", "events", "async", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing",
]

dependencies = [
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "aiohttp>=3.8.0",
    "jsonschema>=4.0.0",
    "aiosqlite>=0.19.0",
    "redis>=4.5.0",
    "aiofiles>=23.0.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
all = [
    "ollama>=0.1.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
]

[project.scripts]
gleitzeit = "gleitzeit.cli.gleitzeit_cli:main"
gz = "gleitzeit.cli.gleitzeit_cli:main"

[project.urls]
"Source" = "https://github.com/leifmarkthaler/gleitzeit"
"Bug Reports" = "https://github.com/leifmarkthaler/gleitzeit/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["gleitzeit*"]

[tool.setuptools.package-data]
gleitzeit = ["protocols/*.yaml", "examples/*.yaml", "examples/*.py", "examples/*.txt", "examples/*.png", "examples/*.md"]